refactor: rename to online_status and create new file
